This fixes an error were a tag was being considered atomic when it wasnt
If the tag 'u' was considered as part of the atomic tags, the tag 'ul'
was being considered as well, even if it wasnt configured to be.

This changes that by adding the possibility to negate values on the tag
(ang only validating once the entire tag is read, not just the first
letters.